[flink] Rename DatabaseSyncMode to MultiTablesSinkMode (#1916)
